added more functionality

+++ musicplayer.py
@@ -2,7 +2,7 @@ import os, time
 import curses
 import subprocess
 import signal
-
+import threading
 
 #ScrapePlayerDESKTOP
 #dependencies:
@@ -10,11 +10,30 @@ import signal
 #		curses
 
 #TODO --
-#-> or <- when no folders, break early
-#hook event on POpen thread finish, play next after finish
 #playlist all in subdir
-#sort by date as well
 #take command line parameters
+#shuffle
+
+#==============begin comparator
+
+def istrcmp(a,b):
+	return cmp(a.lower(),b.lower())
+	
+def songs_by_name_cmp(a,b):
+	global INVERT_SONG
+	return istrcmp(a['name'],b['name']) * INVERT_SONG
+	
+def songs_by_date_cmp(a,b):
+	global INVERT_SONG
+	return cmp(a['ftime'],b['ftime']) * INVERT_SONG
+	
+def folders_by_name_cmp(a,b):
+	global INVERT_FOLDER
+	return istrcmp(a.name,b.name) * INVERT_FOLDER
+	
+SONG_COMPARATOR = songs_by_name_cmp
+INVERT_SONG = 1
+INVERT_FOLDER = 1
 
 #==============begin class definitions
 
@@ -23,8 +42,11 @@ class FolderNode:
 		self.name = name
 		self.songs = {}
 		self.subfolders = {}
-		self.fulldir = "."
+		self.fulldir = "/"
 		self.parent = None
+	
+	def time(self):
+		return os.path.getmtime(self.fulldir)
 		
 	def add_song(self, songdata): 
 		self.songs[songdata['name']] = songdata
@@ -35,13 +57,15 @@ class FolderNode:
 		return self.subfolders[name]
 		
 	def get_songnames(self):
-		rtv = [song for song in self.songs]
-		rtv.sort()
+		rtv = [self.songs[song] for song in self.songs]
+		rtv.sort(SONG_COMPARATOR)
+		rtv = [song['name'] for song in rtv]
 		return rtv
 		
 	def get_foldernames(self):
-		rtv = [folder for folder in self.subfolders]
-		rtv.sort()
+		rtv = [self.subfolders[folder] for folder in self.subfolders]
+		rtv.sort(folders_by_name_cmp)
+		rtv = [folder.name for folder in rtv]
 		return rtv
 		
 class Range2d:
@@ -51,9 +75,20 @@ class Range2d:
 		self.ymin = ymin
 		self.ymax = ymax
 		
+def str_safe_convert(msg):
+	sbuf = ""
+	for i in range(0,len(msg)):
+		try:
+			sbuf = sbuf + str(msg[i])
+		except UnicodeEncodeError:
+			sbuf = sbuf + "?"
+		except KeyError:
+			sbuf = sbuf + "?"
+	return sbuf
+		
 #=============begin folder crawling code
 		
-file_types = ['.mp3','.flac']
+file_types = ['.mp3','.flac','.wav','.ogg','.aiff']
 file_tree = FolderNode("/")
 file_tree.fulldir = os.getcwd()
 current_folder = file_tree
@@ -146,6 +181,7 @@ def draw_ui():
 		current_folder.fulldir
 	))
 	folderbox_range = get_folderbox_range()
+	nc_drawstringat(folderbox_range.xmax-25,folderbox_range.ymin-1,"Last Modified:")
 	for y in range(folderbox_range.ymin,folderbox_range.ymax):
 		for x in range(folderbox_range.xmin,folderbox_range.xmax):
 			nc_drawat(x,y,'+') if y == folderbox_range.ymin or y == folderbox_range.ymax-1 else 0
@@ -155,11 +191,12 @@ def draw_ui():
 		">>>" if current_mode == Mode.SONGS else "   ",
 		"%d/%d"%(songs_offset+1,int( len(current_folder.get_songnames())/ (get_songbox_internal_height()+1) )+1)
 	))
+	nc_drawstringat(songbox_range.xmax-25,songbox_range.ymin-1,"Last Modified:")
 	for y in range(songbox_range.ymin,songbox_range.ymax):
 		for x in range(songbox_range.xmin,songbox_range.xmax): 
 			nc_drawat(x,y,'+') if y == songbox_range.ymin or y == songbox_range.ymax-1 else 0
 			nc_drawat(x,y,'+') if x == songbox_range.xmin or x == songbox_range.xmax-1 else 0
-	nc_drawstringat(0,HEI-5,"(Q)uit (T)pane (Arrows)move (Z/X)pgdown/UP (P)play/pause")
+	nc_drawstringat(0,HEI-5,"(Q)uit (T)pane (Arrows)move (Z/X)pgdown/UP (P)play/pause (o)sort by (I)nvert")
 	nc_drawstringat(0,HEI-4,"=======MESSAGES:=======")
 	
 def draw_folders():
@@ -170,6 +207,7 @@ def draw_folders():
 	for y in range(folderbox_range.ymin+1,folderbox_range.ymax-1):
 		if (i >= len(folders)):
 			break
+		nc_drawstringat(folderbox_range.xmax-26,y,time.ctime(current_folder.subfolders[folders[i]].time()))
 		nc_drawstringat(folderbox_range.xmin+2,y,folders[i])
 		nc_drawstringat(folderbox_range.xmin+1,y,">") if i == folder_localindex + (folder_offset * get_folderbox_internal_height()) else nc_drawstringat(folderbox_range.xmin+1,y," ")  
 		i = i + 1
@@ -182,6 +220,7 @@ def draw_songs():
 	for y in range(songbox_range.ymin+1,songbox_range.ymax-1):
 		if (i >= len(songs)):
 			break
+		nc_drawstringat(songbox_range.xmax-26,y,time.ctime(current_folder.songs[songs[i]]['ftime']))
 		nc_drawstringat(songbox_range.xmin+2,y,songs[i])
 		nc_drawstringat(songbox_range.xmin+1,y,">") if i == songs_localindex + (songs_offset * get_songbox_internal_height()) else nc_drawstringat(songbox_range.xmin+1,y," ") 
 		i = i + 1
@@ -203,6 +242,16 @@ def get_folderbox_internal_height():
 
 #===========BEGIN input loop
 
+def update_screen():
+	clear_screen()
+	draw_ui()
+	nc_drawstringat(0,HEI-3," "*WID)
+	draw_folders()
+	draw_songs()
+	nc_drawstringat(0,HEI-3,debug_output)
+	nc_drawstringat(0,HEI-2,debug_output2)
+	STDSCR.refresh()
+
 def get_actual_songindex():
 	return songs_localindex + songs_offset * get_songbox_internal_height()
 
@@ -215,13 +264,63 @@ def reset_folder_and_song_indexes():
 	folder_offset = 0
 	songs_localindex = 0
 	songs_offset = 0
-
+	
+def song_finished():
+	global songs_localindex, songs_offset, current_folder, currently_playing_poller, currently_playing, is_paused, folder_offset
+	
+	currently_playing = None
+	currently_playing_poller = None
+	use_height = get_songbox_internal_height() if (get_songbox_internal_height() + songs_offset * get_songbox_internal_height()) < len(current_folder.get_songnames()) else len(current_folder.get_songnames()) % get_songbox_internal_height() 
+	ins = songs_localindex
+	
+	songs_localindex = songs_localindex + 1
+	if (songs_localindex >= use_height):
+		songs_localindex = 0
+		songs_offset = songs_offset+1
+		
+	songindex = get_actual_songindex()
+	if (songindex >= len(current_folder.get_songnames())):
+		songs_offset = 0
+	songindex = get_actual_songindex()
+	if (songindex < len(current_folder.get_songnames())):
+		songname = current_folder.get_songnames()[songindex]
+		songdata = current_folder.songs[songname]
+		is_paused = False
+		play_song(songdata["file"])
+		update_screen()
+		
+	else:
+		nc_drawstringat(0,HEI-2,"song_finished but cannot next")
+		STDSCR.refresh()
+	
+	
+def play_song(file):
+	global currently_playing, currently_playing_poller,currently_playing_key,debug_output
+	currently_playing = subprocess.Popen("""play '%s' -q -V0"""%(file.replace("'","\'\\\'\'")),shell=True)
+	currently_playing_poller = ProcessPoller(currently_playing)
+	currently_playing_poller.start()
+	currently_playing_key = file
+	debug_output = """play '%s' -q -V0"""%(file.replace("'","\'\\\'\'"))
+	
 class Mode:
 	FOLDERS = 0
 	SONGS = 1
 	
+class ProcessPoller(threading.Thread):
+	def __init__(self, target):
+		threading.Thread.__init__(self)
+		self.target = target
+		self.kill = False
+		
+	def run(self):
+		self.target.wait()
+		if not self.kill:
+			self.kill = True
+			song_finished()
+	
 current_folder = file_tree
 debug_output = ""
+debug_output2 = ""
 current_mode = Mode.FOLDERS
 folder_localindex = 0
 folder_offset = 0
@@ -229,10 +328,12 @@ songs_localindex = 0
 songs_offset = 0
 
 currently_playing = None
+currently_playing_poller = None
 currently_playing_key = ""
 is_paused = False
 	
 try:
+
 	INPUT = ''
 	while True:		
 		if INPUT == ord('q'):
@@ -240,23 +341,43 @@ try:
 			
 		elif INPUT == ord('t'):
 			current_mode = (current_mode+1)%2
+
+		elif INPUT == ord('i'):
+			if current_mode == Mode.FOLDERS:
+				INVERT_FOLDER = -1 if INVERT_FOLDER == 1 else 1
+				reset_folder_and_song_indexes()
+				debug_output = "invert folder:%d"%INVERT_FOLDER
+							
+			elif current_mode == Mode.SONGS:
+				INVERT_SONG = -1 if INVERT_SONG == 1 else 1
+				reset_folder_and_song_indexes()
+				debug_output = "invert song:%d"%INVERT_SONG
+				
+		elif INPUT == ord('o'):
+			if current_mode == Mode.SONGS:
+				SONG_COMPARATOR = songs_by_name_cmp if SONG_COMPARATOR == songs_by_date_cmp else songs_by_date_cmp
+				debug_output = "song sorted by %s"%("date" if SONG_COMPARATOR == songs_by_date_cmp else "name")
+				reset_folder_and_song_indexes()
+			
 		
 		elif INPUT == curses.KEY_LEFT:
-			reset_folder_and_song_indexes()
-			if (current_folder.parent != None):
-				current_folder = current_folder.parent
-			else:
-				debug_output = "at root"
+			if current_mode == Mode.FOLDERS:
+				reset_folder_and_song_indexes()
+				if (current_folder.parent != None):
+					current_folder = current_folder.parent
+				else:
+					debug_output = "at root"
 				
 			
 		elif INPUT == curses.KEY_RIGHT:
-			if len(current_folder.get_foldernames()) > 0:
-				target_folder = current_folder.get_foldernames()[get_actual_folderindex()]
-				if target_folder in current_folder.subfolders:
-					current_folder = current_folder.subfolders[target_folder]
-					reset_folder_and_song_indexes()
-			else:
-				debug_output = "empty folder" 
+			if current_mode == Mode.FOLDERS:
+				if len(current_folder.get_foldernames()) > 0:
+					target_folder = current_folder.get_foldernames()[get_actual_folderindex()]
+					if target_folder in current_folder.subfolders:
+						current_folder = current_folder.subfolders[target_folder]
+						reset_folder_and_song_indexes()
+				else:
+					debug_output = "empty folder" 
 			
 		elif INPUT == curses.KEY_UP:
 			if current_mode == Mode.FOLDERS:
@@ -304,53 +425,44 @@ try:
 				songs_offset = songs_offset + 1 if (songs_offset+1) * (get_songbox_internal_height()) < len(current_folder.get_songnames()) else 0
 				
 		elif INPUT == ord('p'):
-			if len(current_folder.get_songnames()) > 0:
-				songname = current_folder.get_songnames()[get_actual_songindex()]
-				songdata = current_folder.songs[songname]
+			if current_mode == Mode.SONGS:
+				if len(current_folder.get_songnames()) > 0:
+					songname = current_folder.get_songnames()[get_actual_songindex()]
+					songdata = current_folder.songs[songname]
 		
-				if currently_playing == None:
-					is_paused = False
-					currently_playing = subprocess.Popen("play '%s' -q"%(songdata["file"]),shell=True)
-					currently_playing_key = songdata["file"]
-					debug_output = "now playing: '%s'"%(songdata["file"])
-			
-				else:
-					if songdata["file"] == currently_playing_key:
-						if is_paused:
-							currently_playing.send_signal(signal.SIGCONT)
-							is_paused = False
-							debug_output = "now playing: '%s'"%(songdata["file"])
-						
-						else:
-							currently_playing.send_signal(signal.SIGSTOP)
-							is_paused = True
-							debug_output = "paused: '%s'"%(songdata["file"])
-					
-					else:
+					if currently_playing == None:
 						is_paused = False
-						currently_playing.send_signal(signal.SIGKILL)
-						currently_playing = subprocess.Popen("play '%s' -q"%(songdata["file"]),shell=True)
-						currently_playing_key = songdata["file"]
-						debug_output = "now playing: '%s'"%(songdata["file"])
-				
-			else:
-				debug_output = "no songs in this folder" 
+						play_song(songdata["file"])
+			
+					else:
+						if songdata["file"] == currently_playing_key:
+							if is_paused:
+								currently_playing.send_signal(signal.SIGCONT)
+								is_paused = False
+								debug_output = "now playing: '%s'"%(songdata["file"])
+						
+							else:
+								currently_playing.send_signal(signal.SIGSTOP)
+								is_paused = True
+								debug_output = "paused: '%s'"%(songdata["file"])
+					
+						else:
+							is_paused = False
+							currently_playing_poller.kill = True
+							currently_playing.send_signal(signal.SIGTERM)
+							play_song(songdata["file"])
+
 		
-		clear_screen()
-		draw_ui()
-		nc_drawstringat(0,HEI-3," "*WID)
-		draw_folders()
-		draw_songs()
-		nc_drawstringat(0,HEI-3,debug_output)
-		
-		STDSCR.refresh()
+		update_screen()
 		INPUT = STDSCR.getch()
 except:
 	nc_end()
 	if currently_playing != None:
+		currently_playing_poller.kill = True
 		currently_playing.send_signal(signal.SIGKILL)
 	raise
 finally:
 	if currently_playing != None:
+		currently_playing_poller.kill = True
 		currently_playing.send_signal(signal.SIGKILL)
 	nc_end()
